FT2 async ring: use uint64_t for m_asyncAudioPos to avoid overflow UB
std::atomic<int> overflows after ~50h at 12kHz, which is undefined
behavior and produces negative indices in the reader's modulo calc
— causing out-of-bounds reads from m_asyncAudio.

Switch to std::atomic<uint64_t>, which has well-defined wrap-around
per C++ memory model and keeps the ring-index arithmetic safe.

Simplified the reader start calculation: no longer needs the
'+ ASYNC_BUF_SIZE' workaround since uint64_t subtraction is modular.

+++ DecodiumBridge.h
@@ -1301,7 +1301,10 @@ private:
     // letto dal timer FT2 (onAsyncDecodeTimer). Atomico per garantire una read
     // coerente dello snapshot della write position; il callback usa fetch_add
     // implicito via operator++ per l'incremento.
-    std::atomic<int> m_asyncAudioPos {0};   // posizione assoluta (mai resettata, mod ASYNC_BUF_SIZE)
+    // Tipo unsigned 64-bit: a 12 kHz un int signed va in overflow UB dopo
+    // ~50 h di funzionamento continuo. uint64_t copre >48 milioni di anni e
+    // ha wrap-around well-defined per il C++ memory model.
+    std::atomic<uint64_t> m_asyncAudioPos {0};   // posizione assoluta (mai resettata, mod ASYNC_BUF_SIZE)
     bool  m_asyncDecodePending {false};  // previene overlap
 
     // Spectrum ring buffer — buffer circolare separato per waterfall (non consumato dal decoder)
